move `ManagedAggState` to `state_manager::AggState`
