+++ src/stream/src/executor/aggregation/agg_call.rs
@@ -51,11 +51,10 @@ impl AggArgs {
     }
 }
 
-/// Represents an aggregation function.
+/// Represents an aggregation function call.
 #[derive(Clone, Debug)]
 pub struct AggCall {
-    /// Aggregation Kind for constructing
-    /// [`crate::executor::aggregation::ManagedStateImpl`]
+    /// Aggregation kind for constructing agg state.
     pub kind: AggKind,
     /// Arguments of aggregation function input.
     pub args: AggArgs,
