+++ src/stream/src/executor/aggregation/state_manager.rs
@@ -16,24 +16,148 @@ use std::fmt::Debug;
 
 use itertools::{multizip, Itertools};
 use risingwave_common::array::column::Column;
-use risingwave_common::array::{ArrayBuilderImpl, Op, Row};
+use risingwave_common::array::stream_chunk::Ops;
+use risingwave_common::array::{ArrayBuilderImpl, ArrayImpl, Op, Row};
 use risingwave_common::buffer::Bitmap;
+use risingwave_common::catalog::Schema;
 use risingwave_common::types::Datum;
+use risingwave_expr::expr::AggKind;
+use risingwave_storage::table::streaming_table::state_table::StateTable;
 use risingwave_storage::StateStore;
 
-use super::{AggStateTable, ManagedAggState};
+use super::array_agg::ManagedArrayAggState;
+use super::extreme::GenericExtremeState;
+use super::string_agg::ManagedStringAggState;
+use super::{AggCall, AggStateTable, InMemoryValueState, ManagedTableState};
 use crate::executor::error::StreamExecutorResult;
+use crate::executor::PkIndices;
 
-/// States for [`crate::executor::LocalSimpleAggExecutor`],
+/// Managed state for aggregation. The managed state will manage the cache and/or integrate
+/// the state with the underlying state store.
+enum AggState<S: StateStore> {
+    /// State as single in-memory scalar value, e.g. `count`, `sum`, append-only `min`/`max`.
+    InMemoryValueState(InMemoryValueState),
+
+    /// State as materialized input, e.g. non-append-only `min`/`max`, `string_agg`
+    MaterializedInputState(Box<dyn ManagedTableState<S>>),
+}
+
+impl<S: StateStore> AggState<S> {
+    #[allow(clippy::too_many_arguments)]
+    fn create(
+        agg_call: &AggCall,
+        agg_state_table: Option<&AggStateTable<S>>,
+        row_count: usize,
+        prev_output: Option<&Datum>,
+        pk_indices: &PkIndices,
+        group_key: Option<&Row>,
+        extreme_cache_size: usize,
+        input_schema: &Schema,
+    ) -> StreamExecutorResult<Self> {
+        match agg_call.kind {
+            AggKind::Avg | AggKind::Count | AggKind::Sum | AggKind::ApproxCountDistinct => Ok(
+                Self::InMemoryValueState(InMemoryValueState::new(agg_call, prev_output.cloned())?),
+            ),
+            // optimization: use single-value state for append-only min/max
+            AggKind::Max | AggKind::Min if agg_call.append_only => Ok(Self::InMemoryValueState(
+                InMemoryValueState::new(agg_call, prev_output.cloned())?,
+            )),
+            AggKind::Max | AggKind::Min => Ok(Self::MaterializedInputState(Box::new(
+                GenericExtremeState::new(
+                    agg_call,
+                    group_key,
+                    pk_indices,
+                    agg_state_table
+                        .expect("non-append-only min/max must have state table")
+                        .mapping
+                        .clone(),
+                    row_count,
+                    extreme_cache_size,
+                    input_schema,
+                ),
+            ))),
+            AggKind::StringAgg => Ok(Self::MaterializedInputState(Box::new(
+                ManagedStringAggState::new(
+                    agg_call,
+                    group_key,
+                    pk_indices,
+                    agg_state_table
+                        .expect("string_agg must have state table")
+                        .mapping
+                        .clone(),
+                    row_count,
+                ),
+            ))),
+            AggKind::ArrayAgg => Ok(Self::MaterializedInputState(Box::new(
+                ManagedArrayAggState::new(
+                    agg_call,
+                    group_key,
+                    pk_indices,
+                    agg_state_table
+                        .expect("array_agg must have state table")
+                        .mapping
+                        .clone(),
+                    row_count,
+                ),
+            ))),
+        }
+    }
+
+    async fn apply_chunk(
+        &mut self,
+        ops: Ops<'_>,
+        visibility: Option<&Bitmap>,
+        columns: &[&ArrayImpl],
+        agg_state_table: Option<&mut AggStateTable<S>>,
+    ) -> StreamExecutorResult<()> {
+        match self {
+            Self::InMemoryValueState(state) => state.apply_chunk(ops, visibility, columns),
+            Self::MaterializedInputState(state) => {
+                state
+                    .apply_chunk(
+                        ops,
+                        visibility,
+                        columns,
+                        &mut agg_state_table
+                            .expect("managed table state must have state table")
+                            .table,
+                    )
+                    .await
+            }
+        }
+    }
+
+    async fn get_output(
+        &mut self,
+        agg_state_table: Option<&AggStateTable<S>>,
+    ) -> StreamExecutorResult<Datum> {
+        match self {
+            Self::InMemoryValueState(state) => Ok(state.get_output()),
+            Self::MaterializedInputState(state) => {
+                state
+                    .get_output(
+                        &agg_state_table
+                            .expect("managed table state must have state table")
+                            .table,
+                    )
+                    .await
+            }
+        }
+    }
+}
+
+/// Manager of [`AggState`] for all [`AggCall`]s, used in
 /// [`crate::executor::GlobalSimpleAggExecutor`] and [`crate::executor::HashAggExecutor`].
+///
+/// Agg executors should create an [`AggStateManager`] for each aggregation group.
 pub struct AggStateManager<S: StateStore> {
     /// Group key of the state.
     group_key: Option<Row>,
 
-    /// Current managed states for all [`crate::executor::aggregation::AggCall`]s.
-    managed_states: Vec<ManagedAggState<S>>,
+    /// Managed agg states.
+    states: Vec<AggState<S>>,
 
-    /// Previous outputs of managed states. Initializing with `None`.
+    /// Previous outputs of all managed states. Initialized with `None`.
     ///
     /// We use `Vec<Datum>` instead of `Row` here to avoid unnecessary memory
     /// usage for group key prefix and unnecessary construction of `Row` struct.
@@ -50,7 +174,7 @@ impl<S: StateStore> Debug for AggStateManager<S> {
 }
 
 /// We assume the first state of aggregation is always `StreamingRowCountAgg`.
-pub const ROW_COUNT_COLUMN: usize = 0;
+const ROW_COUNT_COLUMN: usize = 0;
 
 /// Information about the changes built by `AggState::build_changes`.
 pub struct AggChangesInfo {
@@ -63,16 +187,56 @@ pub struct AggChangesInfo {
 }
 
 impl<S: StateStore> AggStateManager<S> {
-    pub fn new(
+    /// Create [`AggStateManager`] for the given [`AggCall`]s.
+    /// For [`crate::executor::HashAggExecutor`], the group key shouldn't be `None`.
+    pub async fn create(
         group_key: Option<Row>,
-        managed_states: Vec<ManagedAggState<S>>,
-        prev_outputs: Option<Vec<Datum>>,
-    ) -> Self {
-        Self {
-            group_key,
-            managed_states,
-            prev_outputs,
+        agg_calls: &[AggCall],
+        agg_state_tables: &[Option<AggStateTable<S>>],
+        result_table: &StateTable<S>,
+        pk_indices: &PkIndices,
+        extreme_cache_size: usize,
+        input_schema: &Schema,
+    ) -> StreamExecutorResult<AggStateManager<S>> {
+        let prev_result: Option<Row> = result_table
+            .get_row(group_key.as_ref().unwrap_or_else(Row::empty))
+            .await?;
+        let prev_outputs: Option<Vec<_>> = prev_result.map(|row| row.0);
+        if let Some(prev_outputs) = prev_outputs.as_ref() {
+            assert_eq!(prev_outputs.len(), agg_calls.len());
         }
+
+        let row_count = prev_outputs
+            .as_ref()
+            .and_then(|outputs| {
+                outputs[ROW_COUNT_COLUMN]
+                    .clone()
+                    .map(|x| x.into_int64() as usize)
+            })
+            .unwrap_or(0);
+
+        let states = agg_calls
+            .iter()
+            .enumerate()
+            .map(|(idx, agg_call)| {
+                AggState::create(
+                    agg_call,
+                    agg_state_tables[idx].as_ref(),
+                    row_count,
+                    prev_outputs.as_ref().map(|outputs| &outputs[idx]),
+                    pk_indices,
+                    group_key.as_ref(),
+                    extreme_cache_size,
+                    input_schema,
+                )
+            })
+            .try_collect()?;
+
+        Ok(Self {
+            group_key,
+            states,
+            prev_outputs,
+        })
     }
 
     pub fn group_key(&self) -> Option<&Row> {
@@ -99,7 +263,7 @@ impl<S: StateStore> AggStateManager<S> {
         // TODO(yuchao): may directly pass `&[Column]` to managed states.
         let column_refs = columns.iter().map(|col| col.array_ref()).collect_vec();
         for (managed_state, agg_state_table, visibility) in
-            multizip((&mut self.managed_states, agg_state_tables, visibilities))
+            multizip((&mut self.states, agg_state_tables, visibilities))
         {
             managed_state
                 .apply_chunk(
@@ -119,7 +283,7 @@ impl<S: StateStore> AggStateManager<S> {
         agg_state_tables: &[Option<AggStateTable<S>>],
     ) -> StreamExecutorResult<Vec<Datum>> {
         futures::future::try_join_all(
-            self.managed_states
+            self.states
                 .iter_mut()
                 .zip_eq(agg_state_tables)
                 .map(|(state, agg_state_table)| state.get_output(agg_state_table.as_ref())),
